修复缺陷2

+++ src/main/java/com/education/hjj/bz/formBean/StudentDemandConnectForm.java
@@ -24,7 +24,7 @@ public class StudentDemandConnectForm extends PageForm implements Serializable {
 	 */
 	private static final long serialVersionUID = 5592231160203780697L;
 
-	private Long sid;
+	private Integer sid;
 
     @ApiModelProperty(value = "需求ID", required = true)
     private Integer demandId;
@@ -172,11 +172,11 @@ public class StudentDemandConnectForm extends PageForm implements Serializable {
         this.createTime = createTime;
     }
 
-    public Long getSid() {
+    public Integer getSid() {
         return sid;
     }
 
-    public void setSid(Long sid) {
+    public void setSid(Integer sid) {
         this.sid = sid;
     }
